+++ src/refactoring/components/admin/ProductManage.tsx
@@ -2,6 +2,7 @@ import { useState } from 'react';
 import { Discount, Product } from '../../../types';
 import { Accordion } from '../ui/Accordion';
 import { useAccordions } from '../../hooks/useAccordions';
+import { AddProductForm } from './AddProductForm';
 
 interface Props {
   products: Product[];
@@ -15,12 +16,6 @@ export const ProductManage = ({ products, onProductUpdate, onProductAdd }: Props
   const [editingProduct, setEditingProduct] = useState<Product | null>(null);
   const [newDiscount, setNewDiscount] = useState<Discount>({ quantity: 0, rate: 0 });
   const [showNewProductForm, setShowNewProductForm] = useState(false);
-  const [newProduct, setNewProduct] = useState<Omit<Product, 'id'>>({
-    name: '',
-    price: 0,
-    stock: 0,
-    discounts: [],
-  });
 
   // handleEditProduct 함수 수정
   const handleEditProduct = (product: Product) => {
@@ -67,8 +62,14 @@ export const ProductManage = ({ products, onProductUpdate, onProductAdd }: Props
         ...updatedProduct,
         discounts: [...updatedProduct.discounts, newDiscount],
       };
+
+      // Entity
       onProductUpdate(newProduct);
+
+      // View
       setEditingProduct(newProduct);
+
+      // Entity
       setNewDiscount({ quantity: 0, rate: 0 });
     }
   };
@@ -85,15 +86,8 @@ export const ProductManage = ({ products, onProductUpdate, onProductAdd }: Props
     }
   };
 
-  const handleAddNewProduct = () => {
-    const productWithId = { ...newProduct, id: Date.now().toString() };
-    onProductAdd(productWithId);
-    setNewProduct({
-      name: '',
-      price: 0,
-      stock: 0,
-      discounts: [],
-    });
+  const handleAddNewProduct = (newProduct: Product) => {
+    onProductAdd(newProduct);
     setShowNewProductForm(false);
   };
 
@@ -108,50 +102,7 @@ export const ProductManage = ({ products, onProductUpdate, onProductAdd }: Props
       </button>
 
       {/* 새 상품 추가 폼 */}
-      {showNewProductForm && (
-        <div className="bg-white p-4 rounded shadow mb-4">
-          <h3 className="text-xl font-semibold mb-2">새 상품 추가</h3>
-          <div className="mb-2">
-            <label htmlFor="productName" className="block text-sm font-medium text-gray-700">
-              상품명
-            </label>
-            <input
-              id="productName"
-              type="text"
-              value={newProduct.name}
-              onChange={(e) => setNewProduct({ ...newProduct, name: e.target.value })}
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <div className="mb-2">
-            <label htmlFor="productPrice" className="block text-sm font-medium text-gray-700">
-              가격
-            </label>
-            <input
-              id="productPrice"
-              type="number"
-              value={newProduct.price}
-              onChange={(e) => setNewProduct({ ...newProduct, price: parseInt(e.target.value) })}
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <div className="mb-2">
-            <label htmlFor="productStock" className="block text-sm font-medium text-gray-700">
-              재고
-            </label>
-            <input
-              id="productStock"
-              type="number"
-              value={newProduct.stock}
-              onChange={(e) => setNewProduct({ ...newProduct, stock: parseInt(e.target.value) })}
-              className="w-full p-2 border rounded"
-            />
-          </div>
-          <button onClick={handleAddNewProduct} className="w-full bg-blue-500 text-white p-2 rounded hover:bg-blue-600">
-            추가
-          </button>
-        </div>
-      )}
+      {showNewProductForm && <AddProductForm addProduct={handleAddNewProduct} />}
 
       <div className="space-y-2">
         {products.map((product, index) => (
